# perf: 'mergefiles' improvements