This commit gives a benchmark report on the use of `[]*ach.File` and `[]ach.File`. It intend to show the improvement in passing by values instead of passing by reference

My benchmark results when using `[]*ach.File`
```
go test . -bench BenchmarkMergeFiles -v -run BenchmarkMergeFiles --benchtime 10s goos:darwin
goos: linux
goarch: amd64
pkg: github.com/moov-io/ach
cpu: 11th Gen Intel(R) Core(TM) i7-11850H @ 2.50GHz
BenchmarkMergeFiles
BenchmarkMergeFiles/MergeFiles
    merge_bench_test.go:148: 1 files merged into 1 files
    merge_bench_test.go:148: 100 files merged into 40 files
    merge_bench_test.go:148: 10000 files merged into 4000 files
    merge_bench_test.go:148: 180822 files merged into 72330 files
BenchmarkMergeFiles/MergeFiles-16                 180822             68727 ns/op            3184 B/op         88 allocs/op
BenchmarkMergeFiles/MergeFiles_3Groups
    merge_bench_test.go:152: 1 files merged into 1 files
    merge_bench_test.go:152: 100 files merged into 55 files
    merge_bench_test.go:152: 10000 files merged into 6005 files
    merge_bench_test.go:152: 76762 files merged into 57046 files
BenchmarkMergeFiles/MergeFiles_3Groups-16          76762            173807 ns/op            8804 B/op        245 allocs/op
BenchmarkMergeFiles/MergeFiles_5Groups
    merge_bench_test.go:155: 1 files merged into 1 files
    merge_bench_test.go:155: 100 files merged into 61 files
    merge_bench_test.go:155: 10000 files merged into 6096 files
    merge_bench_test.go:155: 73902 files merged into 45783 files
BenchmarkMergeFiles/MergeFiles_5Groups-16          73902            159337 ns/op            7928 B/op        220 allocs/op
BenchmarkMergeFiles/MergeFiles_10Groups
    merge_bench_test.go:158: 1 files merged into 1 files
    merge_bench_test.go:158: 100 files merged into 61 files
    merge_bench_test.go:158: 10000 files merged into 6039 files
    merge_bench_test.go:158: 75115 files merged into 51119 files
BenchmarkMergeFiles/MergeFiles_10Groups-16                 75115            165668 ns/op            8356 B/op        232 allocs/op
BenchmarkMergeFiles/MergeFiles_100Groups
    merge_bench_test.go:161: 1 files merged into 1 files
    merge_bench_test.go:161: 100 files merged into 40 files
    merge_bench_test.go:161: 10000 files merged into 6972 files
    merge_bench_test.go:161: 71922 files merged into 43097 files
BenchmarkMergeFiles/MergeFiles_100Groups-16                71922            156825 ns/op            7780 B/op        216 allocs/op
PASS
ok      github.com/moov-io/ach  83.807s
```

Also here are the results when using the `[]ach.File` (without the pointer)
```
go test . -bench BenchmarkMergeFiles -v -run BenchmarkMergeFiles --benchtime 10s goos:darwin
goos: linux
goarch: amd64
pkg: github.com/moov-io/ach
cpu: 11th Gen Intel(R) Core(TM) i7-11850H @ 2.50GHz
BenchmarkMergeFiles
BenchmarkMergeFiles/MergeFiles
    merge_bench_test.go:148: 1 files merged into 1 files
    merge_bench_test.go:148: 100 files merged into 34 files
    merge_bench_test.go:148: 10000 files merged into 3334 files
    merge_bench_test.go:148: 200274 files merged into 66759 files
BenchmarkMergeFiles/MergeFiles-16                 200274             63268 ns/op            3414 B/op         76 allocs/op
BenchmarkMergeFiles/MergeFiles_3Groups
    merge_bench_test.go:152: 1 files merged into 1 files
    merge_bench_test.go:152: 100 files merged into 31 files
    merge_bench_test.go:152: 10000 files merged into 3331 files
    merge_bench_test.go:152: 101288 files merged into 33760 files
BenchmarkMergeFiles/MergeFiles_3Groups-16         101288            121395 ns/op            6574 B/op        150 allocs/op
BenchmarkMergeFiles/MergeFiles_5Groups
    merge_bench_test.go:155: 1 files merged into 1 files
    merge_bench_test.go:155: 100 files merged into 31 files
    merge_bench_test.go:155: 10000 files merged into 3331 files
    merge_bench_test.go:155: 99624 files merged into 33206 files
BenchmarkMergeFiles/MergeFiles_5Groups-16          99624            125300 ns/op            6568 B/op        150 allocs/op
BenchmarkMergeFiles/MergeFiles_10Groups
    merge_bench_test.go:158: 1 files merged into 1 files
    merge_bench_test.go:158: 100 files merged into 31 files
    merge_bench_test.go:158: 10000 files merged into 3329 files
    merge_bench_test.go:158: 93589 files merged into 31191 files
BenchmarkMergeFiles/MergeFiles_10Groups-16                 93589            124098 ns/op            6505 B/op        150 allocs/op
BenchmarkMergeFiles/MergeFiles_100Groups
    merge_bench_test.go:161: 1 files merged into 1 files
    merge_bench_test.go:161: 100 files merged into 34 files
    merge_bench_test.go:161: 10000 files merged into 3301 files
    merge_bench_test.go:161: 96330 files merged into 32010 files
BenchmarkMergeFiles/MergeFiles_100Groups-16                96330            130157 ns/op            6454 B/op        150 allocs/op
PASS
ok      github.com/moov-io/ach  85.414s
```