@adamdecaf 2024-02-07
I put those two results in files so we can compare them with [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat). 

```
$ tail -v -n5 with*txt

==> with-pointer.txt <==
    merge_bench_test.go:161: 71922 files merged into 43097 files
BenchmarkMergeFiles/MergeFiles_100Groups-16                71922            156825 ns/op            7780 B/op        216 allocs/op
PASS
ok      github.com/moov-io/ach  83.807s


==> without-pointer.txt <==
    merge_bench_test.go:161: 96330 files merged into 32010 files
BenchmarkMergeFiles/MergeFiles_100Groups-16                96330            130157 ns/op            6454 B/op        150 allocs/op
PASS
ok      github.com/moov-io/ach  85.414s
```

Running `benchstat` to compare them shows a pretty major decrease in `without-pointer.txt`
```
$ benchstat with-pointer.txt without-pointer.txt 
goos: linux
goarch: amd64
pkg: github.com/moov-io/ach
cpu: 11th Gen Intel(R) Core(TM) i7-11850H @ 2.50GHz
                                   │ with-pointer.txt │          without-pointer.txt          │
                                   │      sec/op      │    sec/op     vs base                 │
MergeFiles/MergeFiles-16                 68.73µ ± ∞ ¹   63.27µ ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_3Groups-16         173.8µ ± ∞ ¹   121.4µ ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_5Groups-16         159.3µ ± ∞ ¹   125.3µ ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_10Groups-16        165.7µ ± ∞ ¹   124.1µ ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_100Groups-16       156.8µ ± ∞ ¹   130.2µ ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                  137.7µ         109.2µ        -20.66%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                   │ with-pointer.txt │          without-pointer.txt           │
                                   │       B/op       │     B/op       vs base                 │
MergeFiles/MergeFiles-16                3.109Ki ± ∞ ¹   3.334Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_3Groups-16        8.598Ki ± ∞ ¹   6.420Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_5Groups-16        7.742Ki ± ∞ ¹   6.414Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_10Groups-16       8.160Ki ± ∞ ¹   6.353Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_100Groups-16      7.598Ki ± ∞ ¹   6.303Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                 6.632Ki         5.598Ki        -15.60%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                   │ with-pointer.txt │         without-pointer.txt          │
                                   │    allocs/op     │  allocs/op   vs base                 │
MergeFiles/MergeFiles-16                  88.00 ± ∞ ¹   76.00 ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_3Groups-16          245.0 ± ∞ ¹   150.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_5Groups-16          220.0 ± ∞ ¹   150.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_10Groups-16         232.0 ± ∞ ¹   150.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
MergeFiles/MergeFiles_100Groups-16        216.0 ± ∞ ¹   150.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                   188.5         130.9        -30.52%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```