@adamdecaf 2024-02-07
Removing pointers from public API methods is a breaking change, so it's harder to merge and release as it will break everyone using the library.