@adamdecaf 2024-02-07
We're talking it over internally to figure out how far we want to break things (if at all). I think we want to try removing pointers entirely from the project if the gains are this large, but that's an even larger change. 

@ckbaum do you have any thoughts?