@midepeter 2024-02-08
Hi @adamdecaf 

Trust you are doing great. Wanted to know if there is an opportunity to work with the moov team or at moov. I really like what is being done here.. I am pretty good with golang, rust and python. I current work part time for a fintech startup also but looking for a more challenging role and all. 

I looked up careers page on moov website but there seem to a US based restriction. Just want to know if there is an opportunity to work for moov. I could start out with maintaining the open source repos with some compensation for a while before integrating with the company fully if possible. 

Or I reach out to @wadearnold directly?

Looking forward to hearing from you.

Thank you
Peter 😊