@adamdecaf 2024-02-12
Yes I'll follow up with you.