@ckbaum 2024-02-12
> We're talking it over internally to figure out how far we want to break things (if at all). I think we want to try removing pointers entirely from the project if the gains are this large, but that's an even larger change.
> 
> @ckbaum do you have any thoughts?

Sorry late to this! Not the end of the world, but yeah we'd be a little grumbly if all the method signatures in the project changed with a version update. Partly because at least for my bank, performance of file-based operations is not a huge priority; ACH is naturally such a slow async process that I doubt we'd notice even a 20% speed improvement of any given operation.

But hey, faster is better 👍 we'd get over it. Especially if there are other breaking changes on the docket.