@adamdecaf 2024-02-12
Thanks for the reply @ckbaum and I agree it may not be very noticeable in a lot of systems. I don't think this PR is enough to  force a breaking change. Perhaps we can remove pointers internally and increase the speed without breaking calling code.