@midepeter 2024-02-13
> Thanks for the reply @ckbaum and I agree it may not be very noticeable in a lot of systems. I don't think this PR is enough to force a breaking change. Perhaps we can remove pointers internally and increase the speed without breaking calling code.

So you mean we should remove the internal pointers and leave the pointers around the calling code for the backward compatibility.