@midepeter 2024-02-16
@adamdecaf 

I have tried to implement it the way you said. I realized the code is tightly coupled across the structure and after i was able to get it. The perfomance changes were nothing to write home about. I think we should probably seek a different manner of improving this which i have started to work on and research about. I will give you heads up if i get any.