@adamdecaf 2024-02-19
Okay thanks. I'm going to close this PR for now since it's a breaking change and we're not ready to adopt that.  I appreciae the help and we can look into using less pointers internally to reduce memory operations.